+++ src/main/kotlin/com/github/jbij/plugins/mt4ij/Bundle.kt
@@ -1,4 +1,4 @@
-package com.github.intellij.plugins.mt4ij
+package com.github.jbij.plugins.mt4ij
 
 import com.intellij.DynamicBundle
 import org.jetbrains.annotations.NonNls