+++ src/main/kotlin/com/github/intellij/plugins/mt4ij/ApiHelpers.kt
@@ -5,6 +5,9 @@ import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.diagnostic.Logger
 import com.intellij.openapi.module.Module
 import com.intellij.openapi.module.ModuleType
+import com.intellij.openapi.progress.ProgressIndicator
+import com.intellij.openapi.progress.ProgressManager
+import com.intellij.openapi.progress.Task
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.roots.ContentEntry
 import com.intellij.openapi.roots.ModifiableRootModel
@@ -25,6 +28,8 @@ import org.jetbrains.jps.model.java.JavaSourceRootType
 class ApiHelpers {
     companion object {
         private val log : Logger = Logger.getInstance(ApiHelpers::class.java)
+
+        private const val SCAN_DELAY = 25L
 
         fun getContentEntry(model: ModifiableRootModel): ContentEntry? {
             val contentEntries = model.contentEntries
@@ -83,41 +88,67 @@ class ApiHelpers {
             val sourcesFolder = "main/${templatesPath}"
             val testsFolder   = "test/${templatesPath}"
 
-            ProjectFileIndex
-                .getInstance(project)
-                .iterateContent { virtualFile ->
-                    val path           = virtualFile.path
-                    val isSourceFolder = path.endsWith(sourcesFolder)
-                    val isTestFolder   = path.endsWith(testsFolder)
+            ProgressManager.getInstance().run(object : Task.Backgroundable(
+                project,
+                Bundle.message("mt4ij.actions.project-scanner.title"),
+            ) {
+                override fun run(indicator: ProgressIndicator) {
+                    indicator.isIndeterminate = false
+                    doScanProject(project, indicator, sourcesFolder, testsFolder)
+                }
+            })
+        }
 
-                    if (isSourceFolder || isTestFolder) {
-                        val model = getModelForFile(project, virtualFile)
+        private fun doScanProject(project: Project, indicator: ProgressIndicator, sourcesFolder: String, testsFolder: String) {
+            val fileIndex      = ProjectFileIndex.getInstance(project)
+            var totalFiles     = 0
+            var processedFiles = 0
 
-                        if (null != model) {
-                            val contentEntry = getContentEntry(model, virtualFile)
+            fileIndex.iterateContent { totalFiles++; true } // Count total files
 
-                            if (null != contentEntry) {
-                                val sourceRoots =
-                                    model.getSourceRoots(if (!isTestFolder) JavaSourceRootType.SOURCE else JavaSourceRootType.TEST_SOURCE)
+            fileIndex.iterateContent { virtualFile ->
+                if (indicator.isCanceled) {
+                    return@iterateContent false
+                }
 
-                                if (!sourceRoots.contains(virtualFile)) {
-                                    contentEntry.addSourceFolder(virtualFile, isTestFolder)
-                                    invokeCommit(model, project)
+                processedFiles++
 
-                                    log.info(
-                                        String.format(
-                                            "Added source folder (after project opening): { virtualFile: \"%s\", isTestFolder: %s }",
-                                            path,
-                                            isTestFolder
-                                        )
-                                    )
-                                }
-                            }
+                indicator.text     = Bundle.message("mt4ij.actions.project-scanner.progress", processedFiles, totalFiles)
+                indicator.text2    = virtualFile.path
+                indicator.fraction = processedFiles.toDouble() / totalFiles
+
+                doScanFile(project, indicator, virtualFile, sourcesFolder, testsFolder)
+                Thread.sleep(SCAN_DELAY)
+
+                true
+            }
+        }
+
+        private fun doScanFile(project: Project, indicator: ProgressIndicator, virtualFile: VirtualFile, sourcesFolder: String, testsFolder: String) {
+            val path           = virtualFile.path
+            val isSourceFolder = path.endsWith(sourcesFolder)
+            val isTestFolder   = path.endsWith(testsFolder)
+
+            if (isSourceFolder || isTestFolder) {
+                val model = getModelForFile(project, virtualFile)
+
+                if (model != null) {
+                    val contentEntry = getContentEntry(model, virtualFile)
+
+                    if (contentEntry != null) {
+                        val sourceRoots =
+                            model.getSourceRoots(if (!isTestFolder) JavaSourceRootType.SOURCE else JavaSourceRootType.TEST_SOURCE)
+
+                        if (!sourceRoots.contains(virtualFile)) {
+                            contentEntry.addSourceFolder(virtualFile, isTestFolder)
+                            invokeCommit(model, project)
+
+                            log.info("Added source folder (after project opening): { virtualFile: \"$path\", isTestFolder: $isTestFolder }")
                         }
                     }
-
-                    true
                 }
+            }
         }
+
     }
 }